fix(rich-text): disable block embeds inside Tables (#813)
